+++ tests/objc/Example.m
@@ -216,4 +216,22 @@ static int _staticIntField = 11;
     return arg1 + arg2;
 }
 
++(struct complex) doStuffWithStruct:(struct simple)simple
+{
+    return (struct complex){
+        .things = {1, 2, 3, 4},
+        .callback = NULL,
+        .s = simple,
+        .next = NULL,
+        .bitfield0 = 0,
+        .bitfield1 = 1,
+        .bitfield2 = 2,
+    };
+}
+
++(struct simple) extractSimpleStruct:(struct complex)complex
+{
+    return complex.s;
+}
+
 @end