+++ CHANGELOG.rst
@@ -6,7 +6,9 @@ Changelog
 
 * Added support for using native Python sequence/mapping syntax with ``NSArray`` and ``NSDictionary``. (jeamland)
 * Added functions for declaring custom conversions between Objective-C type encodings and ``ctypes`` types.
-* Extended the Objective-C type encoding decoder to support block types as well as arbitrary qualifiers and pointers.
+* Added functions for splitting and decoding Objective-C method signature encodings.
+* Added automatic conversion of Python sequences to C arrays or structures in method arguments.
+* Extended the Objective-C type encoding decoder to support block types, bit fields (in structures), typed object pointers, and arbitrary qualifiers. If unknown pointer, array, struct or union types are encountered, they are created and registered on the fly.
 * Changed the ``PyObjectEncoding`` to match the real definition of ``PyObject *``.
 * Fixed the declaration of ``unichar`` (was previously ``c_wchar``, is now ``c_ushort``).
 * Removed the ``get_selector`` function. Use the ``SEL`` constructor instead.